Move components to components directory
