+++ src/components/FontAwesome.js
