+++ src/index.js
@@ -1,6 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import * as Components from '.';
 
-export { default as Menu } from './Menu';
+export { default as Menu } from './components/Menu';
 
 export default Components;
